requirements.txt
fastembed==0.3.6
pandas==2.2.2
python-dotenv>=1.0.1
requests>=2.32.3
gradio>=4.44.0
pillow>=10.4.0